update quit society test
